Added patroling

+++ Assets/Scripts/Behaviour/Based/SubjectParameters.cs
@@ -1,10 +1,11 @@
+using System;
 using UnityEngine;
 
 namespace Behaviour.Based {
 	public abstract class SubjectParameters : MonoBehaviour {
 		public int HealthPoints;
 		public int AttackPoints;
-		public int SpeedPoints;
+		public float SpeedPoints;
 		
 	}
 }
